feat: extension get exporter config when activated
