+++ jupyterlab_pioneer/handlers.py
@@ -4,133 +4,7 @@ import inspect
 import tornado
 from jupyter_server.base.handlers import JupyterHandler
 from jupyter_server.extension.handler import ExtensionHandlerMixin
-from tornado.httpclient import AsyncHTTPClient, HTTPRequest
-from tornado.httputil import HTTPHeaders
-from tornado.escape import to_unicode
 from ._version import __version__
-
-def console_exporter(args: dict) -> dict:
-    """ This exporter sends telemetry data to the browser console.
-
-    Args:
-        args: arguments that would be passed to the exporter function, 
-        defined in the configuration file (except data). 
-        It has the following structure:
-        { \n
-            'id': exporter id, optional, \n
-            'data': telemetry data \n
-        }
-
-    Returns: 
-        dict:
-        { \n
-            'exporter': exporter id or 'ConsoleExporter', \n
-            'message': telemetry data \n
-        }
-    """
-
-    return ({
-        'exporter': args.get('id') or 'ConsoleExporter',
-        'message': args['data']
-    })
-
-def command_line_exporter(args: dict) -> dict:
-    """ This exporter sends telemetry data to the python console jupyter is running on.
-
-    Args:
-        args (dict): arguments that would be passed to the exporter function, 
-        defined in the configuration file (except data). 
-        It has the following structure:
-        { \n
-            'id': exporter id, optional, \n
-            'data': telemetry data \n
-        }
-
-    Returns: 
-        dict:
-        { \n
-            'exporter': exporter id or 'CommandLineExporter', \n
-        }
-    """
-
-    print(args['data'])
-    return ({
-        'exporter': args.get('id') or 'CommandLineExporter',
-    })
-
-def file_exporter(args: dict) -> dict:
-    """ This exporter writes telemetry data to local file.
-
-    Args:
-        args (dict): arguments that would be passed to the exporter function, 
-        defined in the configuration file (except data). 
-        It has the following structure:
-        { \n
-            'id': exporter id, optional, \n
-            'path': path to the target log file, \n
-            'data': telemetry data \n
-        }
-
-    Returns: 
-        dict:
-        { \n
-            'exporter': exporter id or 'FileExporter', \n
-        }
-    """
-
-    with open(args.get('path'), 'a+', encoding='utf-8') as f:
-        json.dump(args['data'], f, ensure_ascii=False, indent=4)
-        f.write(',')
-    return({
-        'exporter': args.get('id') or 'FileExporter',
-    })
-
-async def remote_exporter(args: dict) -> dict:
-    """ This exporter sends telemetry data to a remote http endpoint.
-
-    Args:
-        args (dict): arguments that would be passed to the exporter function, 
-        defined in the configuration file (except data). 
-        It has the following structure:
-        { \n
-            'id': exporter id, optional, \n
-            'url': http endpoint url, \n
-            'params': extra parameters that would be passed to the http endpoint, optional, \n
-            'env': environment variables that would be passed to the http endpoint, optional, \n
-            'data': telemetry data \n
-        }
-
-    Returns:
-        dict:
-        { \n
-            'exporter': exporter id or 'RemoteExporter', \n
-            'message': { \n
-                'code': http response code, \n
-                'reason': http response reason, \n
-                'body': http response body \n
-            } \n
-        }
-    """
-    http_client = AsyncHTTPClient()
-    request = HTTPRequest(
-        url=args.get('url'),
-        method='POST',
-        body=json.dumps({
-            'data': args['data'],
-            'params': args.get('params'), # none if exporter does not contain 'params'
-            'env': [{x: os.getenv(x)} for x in args.get('env')] if (args.get('env')) else []
-        }),
-        headers=HTTPHeaders({'content-type': 'application/json'})
-    )
-    response = await http_client.fetch(request, raise_error=False)
-    return({
-        'exporter': args.get('id') or 'RemoteExporter',
-        'message': {
-            'code': response.code,
-            'reason': response.reason,
-            'body': to_unicode(response.body),
-        },
-    })
 
 class RouteHandler(ExtensionHandlerMixin, JupyterHandler):
 
@@ -149,10 +23,7 @@ class RouteHandler(ExtensionHandlerMixin, JupyterHandler):
             elif resource == 'environ':
                 self.finish(json.dumps(dict(os.environ.items())))
             elif resource == 'config':
-                self.finish(json.dumps({
-                    "activeEvents": self.extensionapp.activeEvents,
-                    "logNotebookContentEvents": self.extensionapp.logNotebookContentEvents 
-                }))
+                self.finish(json.dumps(self.config))
             else:
                 self.set_status(404)
         except Exception as e:
@@ -181,18 +52,20 @@ class RouteHandler(ExtensionHandlerMixin, JupyterHandler):
 
         for each in exporters:
             exporter = each.get('exporter')
-            args = each.get('args') or {} # id, url, path, params, env
-            args['data'] = data
+            active_events = each.get('active_events') or self.extensionapp.active_events
+            if data.get('eventDetail').get('eventName') in active_events:
+                args = each.get('args') or {} # id, url, path, params, env
+                args['data'] = data
 
-            if callable(exporter):
-                if inspect.iscoroutinefunction(exporter):
-                    result = await exporter(args)
+                if callable(exporter):
+                    if inspect.iscoroutinefunction(exporter):
+                        result = await exporter(args)
+                    else:
+                        result = exporter(args)
+                    results.append(result)
                 else:
-                    result = exporter(args)
-                results.append(result)
-            else:
-                results.append({
-                    'message': '[Error] exporter is not callable'
-                })
+                    results.append({
+                        'message': '[Error] exporter is not callable'
+                    })
 
         return results
